refactor: split torrent public facade
